EPMRPP-36949: Code Review fixes - 1
