--- app/src/pages/admin/allUsersPage/allUsersPage.scss
@@ -1,17 +1,0 @@
-.header-buttons {
-  display: flex;
-  justify-content: flex-end;
-  button {
-    margin: 18px 15px 18px 0;
-    &:first-child {
-      border-radius: 2px;
-    }
-    svg {
-      width: 20px;
-      height: 20px;
-    }
-  }
-  @media (max-width: $SCREEN_XS_MAX) {
-    display: none;
-  }
-}
