+++ app/src/pages/admin/allUsersPage/usersToolbar/actionPanel/actionPanel.scss
@@ -1,0 +1,11 @@
+.action-buttons {
+  display: flex;
+  justify-content: flex-end;
+  .action-button {
+    margin: 18px 15px 18px 0;
+    svg {
+      width: 20px;
+      height: 20px;
+    }
+  }
+}
