+++ app/src/pages/admin/allUsersPage/usersToolbar/actionPanel/index.js
@@ -1,0 +1,1 @@
+export { ActionPanel } from './actionPanel';
