+++ app/src/pages/admin/allUsersPage/usersToolbar/index.js
@@ -1,0 +1,1 @@
+export { UsersToolbar } from './usersToolbar';
