+++ app/src/pages/admin/allUsersPage/usersToolbar/usersToolbar.jsx
@@ -1,0 +1,3 @@
+import { ActionPanel } from './actionPanel';
+
+export const UsersToolbar = () => <ActionPanel />;
